Remove old components,pages and functions. Add tailwind. Refactor types
