+++ package.json
@@ -11,6 +11,7 @@
     "@types/node": "^16.18.31",
     "@types/react": "^18.2.6",
     "@types/react-dom": "^18.2.4",
+    "assets": "file:./src/assets",
     "bootstrap": "^5.2.3",
     "firebase": "^9.22.0",
     "react": "^18.2.0",
@@ -22,9 +23,8 @@
     "react-tabs": "^6.0.1",
     "react-toastify": "^9.1.3",
     "types": "file:./types",
-    "assets": "file:./src/assets",
-    "utils": "file:./utils",
     "typescript": "^4.9.5",
+    "utils": "file:./utils",
     "web-vitals": "^2.1.4"
   },
   "scripts": {
@@ -50,5 +50,8 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
+  },
+  "devDependencies": {
+    "tailwindcss": "^3.4.1"
   }
 }
