+++ src/App.tsx
@@ -7,13 +7,9 @@ import Loading from "./components/Loading";
 import Navbar from "./components/navbar";
 
 import Lobby from "./pages/lobby";
-import GameBoard from "./pages/game-board";
-import GameParty from "./pages/game-party";
+
 import Home from "./pages/home";
-import Leaderboard from "./pages/leaderboard";
-import Test from "./pages/Test";
-import Tutorial from "./pages/tutorial";
-import GameResult from "./pages/game-result";
+
 import NotFound from "./pages/not-founds";
 import { useAuth } from "./context/AuthContext";
 
@@ -27,13 +23,9 @@ function App() {
       <Navbar />
       <Routes>
         <Route path="/" element={<Home />} />
-        <Route path="/leaderboard" element={<Leaderboard />} />
+
         <Route path="/lobby" element={<Lobby />} />
-        <Route path="/tutorial" element={<Tutorial />} />
-        <Route path="/party/:roomID" element={<GameParty />} />
-        <Route path="/gameboard/:roomID" element={<GameBoard />} />
-        <Route path="/result/:roomID" element={<GameResult />} />
-        <Route path="/test" element={<Test />} />
+
         <Route path="*" element={<NotFound />} />
       </Routes>
     </div>
