--- src/pages/Test.tsx
@@ -1,9 +1,0 @@
-import React from "react";
-
-type Props = {};
-
-const Test = (props: Props) => {
-  return <div>Test</div>;
-};
-
-export default Test;
