--- src/pages/game-board.css
@@ -1,35 +1,0 @@
-.game-component {
-  height: 100%;
-  width: 100%;
-  display: flex;
-  margin: 0 0 1.5rem 0;
-}
-
-.game-component .left {
-  margin: 0 1rem 0 1.5rem;
-  height: 100%;
-  width: 70%;
-}
-
-.game-component .right {
-  display: flex;
-  flex-direction: column;
-  height: 100%;
-  width: 30%;
-}
-
-.game-component .right .top {
-  background-color: rgba(0, 0, 0, 0.1);
-  padding: 1rem;
-  margin: 0 1rem 0rem 0;
-  height: 100%;
-  border-radius: 5px;
-}
-
-.game-component .right .bottom {
-  /* background-color: rgba(0, 0, 0, 0.1); */
-  display: flex;
-  margin: 0 1rem 0 0;
-  height: 30%;
-  border-radius: 5px;
-}
