--- src/pages/game-party.css
@@ -1,36 +1,0 @@
-.game-room-left {
-  background-color: rgba(0, 0, 0, 0.1);
-  padding: 3rem 1rem 0 1rem;
-  margin: 0 1rem 0 2rem;
-  height: 100%;
-  width: 20%;
-}
-
-.game-room-middle {
-  background-color: rgba(0, 0, 0, 0.1);
-  padding: 3rem 1rem 0 1rem;
-  margin: 0 1rem 0 1rem;
-  height: 100%;
-  width: 30%;
-}
-
-.game-room-right {
-  background-color: rgba(0, 0, 0, 0.1);
-  padding: 3rem 1rem 0 1rem;
-  margin: 0 2rem 0 1rem;
-  height: 100%;
-  width: 30%;
-}
-
-.game-room-left .tab {
-  background-color: rgba(0, 0, 0, 0.1);
-  margin: 0 0 0.5rem 0;
-  width: 100%;
-  height: 100%;
-  padding: 0.8rem 0.5rem 0.8rem 0.5rem;
-  font-size: 0.9rem;
-}
-
-.game-room-left .tab:hover {
-  background-color: rgba(0, 0, 0, 0.2);
-}
