--- src/pages/leaderboard.tsx
@@ -1,9 +1,0 @@
-const Leaderboard = () => {
-  return (
-    <div className="w-100 h-100 d-flex justify-content-center align-items-center">
-      Leaderboard(Coming Soon!)
-    </div>
-  );
-};
-
-export default Leaderboard;
