--- src/pages/tutorial.css
@@ -1,24 +1,0 @@
-.tutorial-container {
-  padding: 20px;
-  background-color: rgba(0, 0, 0, 0.2);
-  display: flex;
-  flex-direction: column;
-  overflow-y: scroll;
-  height: 80%;
-}
-
-.tutorial-title {
-  text-align: center;
-  font-size: 24px;
-  margin-bottom: 20px;
-}
-
-.section-title {
-  font-size: 20px;
-  margin-top: 3rem;
-}
-
-.section-description {
-  font-size: 16px;
-  margin-bottom: 15px;
-}
