+++ types/Avatar.ts
@@ -1,0 +1,1 @@
+export type AvatarID = "blueDino" | "redDino" | "yellowDino" | "greenDino";
