+++ types/Card.ts
@@ -1,24 +1,8 @@
-export type Suit = "♣" | "♦" | "♥" | "♠";
+export type Suit = "♠" | "♣" | "♥" | "♦";
 
-export type CardStringValue =
-  | "A"
-  | "K"
-  | "Q"
-  | "J"
-  | "10"
-  | "9"
-  | "8"
-  | "7"
-  | "6"
-  | "5"
-  | "4"
-  | "3"
-  | "2";
+export type Rank = "2" | "3" | "4" | "5" | "6" | "7" | "8" | "9" | "10" | "J" | "Q" | "K" | "A";
 
-export type CardValue = 2 | 3 | 4 | 5 | 6 | 7 | 8 | 9 | 10 | 11 | 12 | 13 | 14;
-
-export type Card = {
+export interface Card {
   suit: Suit;
-  value: CardValue;
-  stringValue: CardStringValue;
-};
+  rank: Rank;
+}
