--- types/Chat.ts
@@ -1,10 +1,0 @@
-import { Timestamp } from "firebase/firestore";
-
-export interface Message {
-  // Player info
-  uid: string;
-  playerName: string;
-
-  text: string;
-  createdAt: Timestamp;
-}
